+++ Project/Source/Modules/ModuleRender.cpp
@@ -50,11 +50,11 @@ static std::vector<float> mediumGaussKernel;
 static std::vector<float> largeGaussKernel;
 
 float defIntGaussian(const float x, const float mu, const float sigma) {
-	return 0.5 * erf((x - mu) / (sqrt(2) * sigma));
+	return (float) (0.5f * erf((x - mu) / (sqrt(2) * sigma)));
 }
 
 void gaussianKernel(const int kernelSize, const float sigma, const float mu, const float step, std::vector<float>& coeff) {
-	const float end = 0.5*kernelSize;
+	const float end = 0.5f*kernelSize;
 	const float start = -end;
 	float sum = 0;
 	float x = start;
@@ -71,7 +71,7 @@ void gaussianKernel(const int kernelSize, const float sigma, const float mu, con
 
 	//normalize
 	sum = 1 / sum;
-	for (int i = 0; i < coeff.size(); ++i) {
+	for (unsigned int i = 0u; i < coeff.size(); ++i) {
 		coeff[i] *= sum;
 	}
 
@@ -706,7 +706,7 @@ UpdateStatus ModuleRender::Update() {
 		glBindTexture(GL_TEXTURE_2D, colorTextures[1]);
 		glGenerateMipmap(GL_TEXTURE_2D);
 
-		for (unsigned int i = 0; i < 2 * bloomQuality; i++) {
+		for (unsigned int i = 0u; i < 2u * bloomQuality; i++) {
 			int width = static_cast<int>(viewportSize.x);
 			int height = static_cast<int>(viewportSize.y);
 			glViewport(0, 0, width / (1 << gaussSmallMipLevel), height / (1 << gaussSmallMipLevel));
@@ -973,10 +973,10 @@ void ModuleRender::UpdateFramebuffers() {
 	}
 
 	// Compute Gaussian kernels
-	gaussSmallKernelRadius = roundf(viewportSize.y * 0.002f);
-	gaussMediumKernelRadius = roundf(viewportSize.y * 0.004f);
-	gaussLargeKernelRadius = roundf(viewportSize.y * 0.008f);
-	float term = Ln(1e5 / sqrt(2 * pi));
+	gaussSmallKernelRadius = (int) roundf(viewportSize.y * 0.002f);
+	gaussMediumKernelRadius = (int) roundf(viewportSize.y * 0.004f);
+	gaussLargeKernelRadius = (int) roundf(viewportSize.y * 0.008f);
+	float term = Ln(1e5f / sqrt(2 * pi));
 	float sigma1 = gaussSmallKernelRadius * gaussSmallKernelRadius / 2.0f;
 	float sigma2 = gaussMediumKernelRadius * gaussMediumKernelRadius / 2.0f;
 	float sigma3 = gaussLargeKernelRadius * gaussLargeKernelRadius / 2.0f;
